change ID type from int to int64

+++ internal/types/types.go
@@ -1,7 +1,7 @@
 package types
 
 type Student struct {
-	ID    int
+	ID    int64
 	Name  string `validate:"required,min=5,max=100"`
 	Email string `validate:"required,email"`
 	RollNo int `validate:"required"`
